Binary Search Tree
The class Node creates a new node with a given data and left and right properties that are initially set to null. The class BinarySearchTree has methods to insert a new node, search for a node, and traverse the tree using inOrderTraverse, preOrderTraverse, and postOrderTraverse methods.

The insert method takes in a new data and creates a new Node instance. If the root is null, the new node becomes the root of the tree. Otherwise, the insertNode method is called to insert the new node in the appropriate place in the tree.

The search method takes in a node and a data, and returns the node with the given data. If the node is null, it returns null. If the data is less than the node's data, it recursively searches the left subtree. If the data is greater than the node's data, it recursively searches the right subtree. If the data is equal to the node's data, it returns the node.

The three traverse methods (inOrderTraverse, preOrderTraverse, postOrderTraverse) take in a node and a callback function, and traverse the tree in-order, pre-order, or post-order, respectively. In each case, the callback function is called on the node.data for each node in the tree.

Finally, the code creates a new instance of BinarySearchTree, inserts some data into the tree, and outputs the result of the in-order, pre-order, and post-order traversals, as well as the result of searching for the node with data equal to 5.