HashTable
This code defines a HashTable class in JavaScript with methods to add new key-value pairs, retrieve the value associated with a given key, retrieve an array of all keys in the hash table, and retrieve an array of all values in the hash table. The hash table uses a hash function to convert the keys into array indices, and uses an array of arrays to store the key-value pairs. This allows for constant time O(1) lookups of values given a key.