CRUD Operations with PostgreSQL Database
This code uses the pg library to connect to a PostgreSQL database and perform CRUD operations on a users table. It defines functions to create the table (if it doesn't already exist), add a new user, retrieve all users, update an existing user, and delete a user. These functions are then used in a series of .then() calls that execute the operations one after the other, logging the results to the console. Finally, it ends the client connection.